[Code] Apply PVS-Studio bugfixes and suggestions, 2022-09

+++ src/_P129_74HC165.ino
@@ -296,7 +296,7 @@ boolean Plugin_129(uint8_t function, struct EventStruct *event, String& string)
         }
 
         for (uint8_t j = 0; j < 8; j++) {
-          bitWrite(bits, off * 8 + (7 - j), isFormItemChecked(getPluginCustomArgName((i * 8 + (7 - j)) + 1)));
+          bitWrite(bits, static_cast<uint64_t>(off * 8 + (7 - j)), isFormItemChecked(getPluginCustomArgName((i * 8 + (7 - j)) + 1)));
         }
         PCONFIG_ULONG(i / 4) = bits;
